flask/requirements.txt
Flask
gunicorn
PyYaml
google-auth
google-cloud-firestore
google-cloud-secret-manager
pandas
plotly
